requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
hdbscan==0.8.29
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2
rich==13.5.2
tqdm==4.65.0
umap-learn==0.5.3
networkx==3.1
python-dateutil==2.8.2
pytz==2023.3
jellyfish==1.2.0 